inference/segmentation/requirements.txt
albumentations==1.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
cmake==3.26.3
efficientnet-pytorch==0.7.1
filelock==3.12.0
fsspec==2023.5.0
huggingface-hub==0.14.1
idna==3.4
imageio==2.29.0
Jinja2==3.1.2
joblib==1.2.0
lazy_loader==0.2
lit==16.0.5
MarkupSafe==2.1.2
mpmath==1.3.0
munch==3.0.0
networkx==3.1
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
opencv-python-headless==4.7.0.72
packaging==23.1
patchify==0.2.3
Pillow==9.5.0
pretrainedmodels==0.7.4
PyWavelets==1.4.1
PyYAML==6.0
qudida==0.0.4
requests==2.31.0
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.10.1
segmentation-models-pytorch==0.3.2
six==1.16.0
sympy==1.12
threadpoolctl==3.1.0
tifffile==2023.4.12
timm==0.6.12
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
triton==2.0.0
typing_extensions==4.6.2
urllib3==2.0.2
tabulate==0.9.0
matplotlib